fix: use double instead of int (might cause losses in accuracy)

+++ fintech/src/main/java/ru/kslacker/fintech/models/Weather.java
@@ -13,6 +13,6 @@ import java.util.UUID;
 public class Weather {
     private UUID regionId;
     private String regionName;
-    private int temperatureValue;
+    private double temperatureValue;
     private LocalDateTime dateTime;
 }
